create service section and news section , board of dirctors

+++ app/Http/Controllers/BoardOfDirctorsController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Resources\BoardOfDirctorsResource;
 use App\Models\BoardOfDirctor;
 use Illuminate\Http\Request;
+
 
 class BoardOfDirctorsController extends Controller
 {
@@ -16,7 +18,7 @@ class BoardOfDirctorsController extends Controller
         return response()->json([
             'status' => true,
             'message' => 'Board of directors retrieved successfully',
-            'data' => $boardOfDirctors
+            'data' => BoardOfDirctorsResource::collection($boardOfDirctors)
         ], 200);
     }
 
@@ -47,49 +49,70 @@ class BoardOfDirctorsController extends Controller
         return response()->json([
             'status' => true,
             'message' => 'Board of director created successfully',
-            'data' => $boardOfDirctor
+            'data' => new BoardOfDirctorsResource($boardOfDirctor)
         ], 201);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(BoardOfDirctor $boardOfDirctor)
+    public function show($id)
     {
+        $boardOfDirctor = BoardOfDirctor::find($id);
+        if (!$boardOfDirctor) {
+            return response()->json([
+                'status' => false,
+                'message' => 'Board of director not found',
+            ], 404);
+        };
         return response()->json([
             'status' => true,
             'message' => 'Board of director retrieved successfully',
-            'data' => $boardOfDirctor
+            'data' => new BoardOfDirctorsResource($boardOfDirctor)
         ], 200);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, BoardOfDirctor $boardOfDirctor)
+    public function update(Request $request, $id)
     {
+        $boardOfDirctor = BoardOfDirctor::find($id);
+        if (!$boardOfDirctor) {
+            return response()->json([
+                'status' => false,
+                'message' => 'Board of director not found',
+            ], 404);
+        };
         $request->validate([
-            'name' => 'required|string',
-            'position' => 'required|string',
-            'image_path' => 'required|string',
-            'description' => 'required|string',
-            'email' => 'required|email',
-            'phone' => 'required|string',
+            'name' => 'nullable|string',
+            'position' => 'nullable|string',
+            'image_path' => 'nullable|string',
+            'description' => 'nullable|string',
+            'email' => 'nullable|email',
+            'phone' => 'nullable|string',
         ]);
 
         $boardOfDirctor->update($request->all());
         return response()->json([
             'status' => true,
             'message' => 'Board of director updated successfully',
-            'data' => $boardOfDirctor
+            'data' => new BoardOfDirctorsResource($boardOfDirctor)
         ], 200);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(BoardOfDirctor $boardOfDirctor)
+    public function destroy($id)
     {
+        $boardOfDirctor = BoardOfDirctor::find($id);
+        if (!$boardOfDirctor) {
+            return response()->json([
+                'status' => false,
+                'message' => 'Board of director not found',
+            ], 404);
+        };
         $boardOfDirctor->delete();
         return response()->json([
             'status' => true,
